requirements.txt
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.1
django-cors-headers==2.4.0
django-filter==2.0.0
django-redis==4.10.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-jwt-custom-user==1.11.1
gunicorn==19.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysqlclient==1.3.13
openapi-codec==1.3.2
Pillow==5.4.1
PyJWT==1.6.4
pytz==2018.7
redis==3.0.1
requests==2.20.1
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
